Fix/Tidy

+++ wangle-server/src/convert_str.hpp
@@ -3,7 +3,7 @@
 
 template<typename Int>
 Int a2n(const char* s){
-	uint64_t n = 0;
+	Int n = 0;
 	while(*s >= '0'  &&  *s <= '9'){
 		n *= 10;
 		n += *s - '0';
@@ -14,7 +14,7 @@ Int a2n(const char* s){
 
 template<typename Int>
 Int a2n(const char** s){
-	uint64_t n = 0;
+	Int n = 0;
 	while(**s >= '0'  &&  **s <= '9'){
 		n *= 10;
 		n += **s - '0';
